requirements.txt
absl-py==1.3.0
aiohttp==3.8.1
aiosignal==1.2.0
alembic==1.8.1
appdirs==1.4.4
arabic-reshaper==3.0.0
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asgiref==3.5.2
asn1crypto==1.5.1
asttokens==2.0.5
astunparse==1.6.3
async-timeout==4.0.2
attrs==21.4.0
backcall==0.2.0
banal==1.0.6
beautifulsoup4==4.10.0
bleach==4.1.0
branca==0.5.0
cachetools==5.3.1
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.9
click==8.0.3
colorama==0.4.4
comm==0.1.4
cryptography==39.0.1
cssselect2==0.7.0
cycler==0.11.0
dataset==1.5.2
debugpy==1.8.0
decorator==5.1.1
defusedxml==0.7.1
distlib==0.3.4
dj-database-url==1.2.0
Django==3.0.5
django-crispy-forms==1.14.0
django-environ==0.9.0
django-rest-framework==0.1.0
django-widget-tweaks==1.4.12
django-xhtml2pdf==0.0.4
django_debug_toolbar==3.8.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
djongo==1.3.6
dnspython==2.4.2
entrypoints==0.4
environ==1.0
et-xmlfile==1.1.0
executing==0.8.3
filelock==3.4.2
Flask==2.0.2
flatbuffers==23.5.26
folium==0.12.1.post1
fontawesomefree==6.2.0
fonttools==4.31.1
frozenlist==1.3.1
future==0.18.2
gast==0.5.4
geocoder==1.38.1
google-ai-generativelanguage==0.3.3
google-api-core==2.14.0
google-auth==2.23.2
google-auth-oauthlib==1.0.0
google-generativeai==0.2.2
google-pasta==0.2.0
googleapis-common-protos==1.61.0
greenlet==1.1.3.post0
grpcio==1.59.3
grpcio-status==1.59.3
h5py==3.10.0
html5lib==1.1
idna==3.3
importlib-metadata==6.8.0
ipykernel==6.25.2
ipython==8.1.1
ipython-genutils==0.2.0
ipywidgets==8.1.1
itsdangerous==2.0.1
jedi==0.18.1
Jinja2==3.0.3
joblib==1.2.0
jsonschema==4.4.0
jupyter==1.0.0
jupyter-client==7.1.2
jupyter-console==6.6.3
jupyter_core==5.3.2
jupyterlab-pygments==0.1.2
jupyterlab-widgets==3.0.9
keras==2.14.0
kiwisolver==1.4.0
libclang==16.0.6
lxml==4.9.2
Mako==1.2.3
Markdown==3.5
MarkupSafe==2.0.1
matplotlib==3.5.1
matplotlib-inline==0.1.3
mediapipe==0.8.11
mistune==0.8.4
ml-dtypes==0.2.0
multidict==6.0.2
mysqlclient==2.1.1
nbclient==0.5.13
nbconvert==6.4.4
nbformat==5.2.0
nest-asyncio==1.5.4
notebook==6.4.10
numpy==1.21.4
oauthlib==3.2.2
openai==0.11.3
opencv-contrib-python==4.6.0.66
opencv-python==4.6.0.66
openpyxl==3.0.9
opt-einsum==3.3.0
oscrypto==1.3.0
packaging==21.3
pandas==1.3.4
pandas-stubs==1.2.0.39
pandocfilters==1.5.0
parso==0.8.3
pickleshare==0.7.5
Pillow==9.0.1
pipenv==2022.1.8
platformdirs==3.11.0
prometheus-client==0.13.1
prompt-toolkit==3.0.39
proto-plus==1.22.3
protobuf==4.25.1
psutil==5.9.0
psycopg2==2.9.5
pure-eval==0.2.2
pyaes==1.6.1
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
pycryptodomex==3.15.0
pyDes==2.0.1
pyee==8.2.2
Pygments==2.11.2
pyHanko==0.17.0
pyhanko-certvalidator==0.20.0
PyJWT==2.6.0
pymongo==3.11.3
pynput==1.7.6
pyparsing==3.0.7
pypdf==3.4.1
pypng==0.20220715.0
pyppeteer==1.0.2
pyrsistent==0.18.1
python-bidi==0.4.2
python-dateutil==2.8.2
python-decouple==3.6
python-dotenv==1.0.0
pytz==2021.3
pytz-deprecation-shim==0.1.0.post0
pywin32==303
pywinpty==2.0.5
PyYAML==6.0
pyzmq==22.3.0
qrcode==7.4.2
qtconsole==5.4.4
QtPy==2.4.0
ratelim==0.1.6
regex==2022.10.31
reportlab==3.6.12
requests==2.26.0
requests-oauthlib==1.3.1
rsa==4.9
scikit-learn==1.1.2
scipy==1.9.2
Send2Trash==1.8.0
six==1.16.0
sklearn==0.0
soupsieve==2.3.1
SQLAlchemy==1.4.42
sqlparse==0.2.4
stack-data==0.2.0
stopwords==1.0.0
stripe==4.1.0
svglib==1.5.1
tensorboard==2.14.1
tensorboard-data-server==0.7.1
tensorflow==2.14.0
tensorflow-estimator==2.14.0
tensorflow-intel==2.14.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.3.0
terminado==0.13.3
testpath==0.6.0
textblob==0.17.1
threadpoolctl==3.1.0
tinycss2==1.2.1
tornado==6.1
tqdm==4.62.3
traitlets==5.11.2
tweepy==4.12.1
typing_extensions==4.5.0
tzdata==2021.5
tzlocal==4.2
uritools==4.0.1
urllib3==1.26.7
vboxapi==1.0
virtualenv==20.13.0
virtualenv-clone==0.5.7
wcwidth==0.2.5
webencodings==0.5.1
websockets==10.4
Werkzeug==2.0.2
whitenoise==6.3.0
widgetsnbextension==4.0.9
wikipedia==1.4.0
wordcloud==1.8.2.2
wrapt==1.14.1
xhtml2pdf==0.2.9
yarl==1.8.1
zipp==3.17.0
